requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
astroid==2.5.1; python_version >= '3.6'
django-rest-auth==0.9.5
django==3.1.7
djangorestframework==3.12.2
isort==5.7.0; python_version >= '3.6' and python_version < '4.0'
lazy-object-proxy==1.5.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mccabe==0.6.1
pylint-django==2.4.2
pylint-plugin-utils==0.6
pylint==2.7.2
pytz==2021.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
wrapt==1.12.1
